+++ engine/api/src/main/scala/pl/touk/nussknacker/engine/api/process/ExpressionConfig.scala
@@ -3,13 +3,15 @@ package pl.touk.nussknacker.engine.api.process
 import pl.touk.nussknacker.engine.api.SpelExpressionExcludeList
 import pl.touk.nussknacker.engine.api.dict.DictDefinition
 import pl.touk.nussknacker.engine.api.expression.ExpressionParser
+import pl.touk.nussknacker.engine.api.process.ExpressionConfig.defaultAdditionalClasses
 
+import java.time.{Duration, LocalDate, LocalDateTime, LocalTime, Period}
 import scala.util.matching.Regex
 
 //TODO: how to make this config less spel-centric?, move globalImports and optimizeCompilation to spel configuration
 case class ExpressionConfig(globalProcessVariables: Map[String, WithCategories[AnyRef]],
                             globalImports: List[WithCategories[String]],
-                            additionalClasses: List[Class[_]] = List.empty,
+                            additionalClasses: List[Class[_]] = defaultAdditionalClasses,
                             languages: LanguageConfiguration = LanguageConfiguration.default,
                             optimizeCompilation: Boolean = true,
                             strictTypeChecking: Boolean = true,
@@ -27,6 +29,11 @@ case class ExpressionConfig(globalProcessVariables: Map[String, WithCategories[A
 
 object ExpressionConfig {
   val empty = ExpressionConfig(Map.empty, Nil)
+
+  // Those types must be explicitly provided because can be used in dynamic parameters
+  val standardEditorClasses: List[Class[_]] = List(classOf[LocalDate], classOf[LocalTime], classOf[LocalDateTime], classOf[Duration], classOf[Period])
+
+  val defaultAdditionalClasses: List[Class[_]] = standardEditorClasses
 }
 
 object LanguageConfiguration {